104968: Move the job file, update specs, remove unnecessary requires
