+++ config/features.yml
@@ -1916,7 +1916,7 @@ features:
   pension_form_profile_module_enabled:
     actor_type: user
     description: Use the module version of the FormProfile
-  pension_event_bus_submission_enabled:
+  pension_kafka_event_bus_submission_enabled:
     actor_type: user
     description: Enable the EventBusSubmissionJob for Kafka
   income_and_assets_form_enabled:
