+++ modules/pensions/spec/e2e_spec.rb
@@ -31,6 +31,7 @@ RSpec.describe 'Pensions End to End', type: :request do
     expect(SavedClaimSerializer).to receive(:new).and_call_original
     expect(PersistentAttachment).to receive(:where).with(guid: anything).and_call_original
     expect(Pensions::PensionBenefitIntakeJob).to receive(:perform_async)
+    expect(Kafka::EventBusSubmissionJob).to receive(:perform_async).twice.and_call_original
     expect(monitor).to receive(:track_create_success).and_call_original
 
     post '/pensions/v0/claims', params: { param_name => { form: form.form } }
